feat: add various transactions
