+++ packages/sdk/src/Configuration/Fees/Exit.ts
@@ -87,6 +87,21 @@ export function decodeDirectFeeSettings(settings: Hex): DirectFeeSettings {
   return { inKindRateInBps, specificAssetsRate, recipient };
 }
 
+export type SetRecipientParams = {
+  comptrollerProxy: Address;
+  fee: Address;
+  recipient: Address;
+};
+
+export function setRecipient(args: SetRecipientParams) {
+  return new Viem.PopulatedTransaction({
+    abi: Abis.IExitRateDirectFee,
+    functionName: "setRecipientForFund",
+    address: args.fee,
+    args: [args.comptrollerProxy, args.recipient],
+  });
+}
+
 //--------------------------------------------------------------------------------------------
 // READ - BOTH TYPES
 //--------------------------------------------------------------------------------------------
