+++ packages/sdk/src/Relayer.ts
@@ -2,6 +2,26 @@ import * as Abis from "@enzymefinance/abis";
 import { type Address, type PublicClient, isAddressEqual, zeroAddress } from "viem";
 import { Viem } from "./Utils.js";
 
+//--------------------------------------------------------------------------------------------
+// TRANSACTIONS
+//--------------------------------------------------------------------------------------------
+
+export type DeployGasRelayPaymasterParams = {
+  comptrollerProxy: Address;
+};
+
+export function deployGasRelayPaymaster(args: DeployGasRelayPaymasterParams) {
+  return new Viem.PopulatedTransaction({
+    abi: Abis.IComptrollerLib,
+    functionName: "deployGasRelayPaymaster",
+    address: args.comptrollerProxy,
+  });
+}
+
+//--------------------------------------------------------------------------------------------
+// READ FUNCTIONS
+//--------------------------------------------------------------------------------------------
+
 export async function isRelayerEnabled(
   client: PublicClient,
   args: Viem.ContractCallParameters<{
